Refactor adfer review
